pulling out helpful functions into a util
